src/main/resources/archetype-resources/__rootArtifactId__.core/src/main/java/core/internal/SampleDelegateCommandHandler.java
#set( $symbol_pound = '#' )
#set( $symbol_dollar = '$' )
#set( $symbol_escape = '\' )
package ${package}.core.internal;


import java.util.List;

import org.eclipse.core.runtime.IProgressMonitor;
import org.eclipse.jdt.ls.core.internal.IDelegateCommandHandler;

public class SampleDelegateCommandHandler implements IDelegateCommandHandler {

    public static final String COMMAND_ID = "${package}.samplecommand";

    @Override
    public Object executeCommand(String commandId, List<Object> arguments, IProgressMonitor progress) throws Exception {
        if (COMMAND_ID.equals(commandId)) {
            return "Hello World";
        }
        throw new UnsupportedOperationException(String.format("Unsupported command '%s'!", commandId));
    }


}
